Split index and SQL extractors into partial files
