+++ src/CodeIndex/Indexer/References/Languages/SqlReferenceExtractor.cs
@@ -4,23 +4,9 @@ using CodeIndex.Models;
 
 namespace CodeIndex.Indexer;
 
-internal static class SqlReferenceExtractor
+internal static partial class SqlReferenceExtractor
 {
-    internal readonly record struct DefinitionLeafSpan(string LeafName, int StartIndex, int EndIndexExclusive);
-
     private readonly record struct CteBodySpan(int StartIndex, int EndIndexExclusive);
-
-    internal readonly record struct IdentifierScanState(
-        bool InBlockComment,
-        string? DollarQuoteDelimiter,
-        bool InSingleQuotedString);
-
-    internal sealed class State
-    {
-        public HashSet<string> EstablishedTempObjectNames { get; } = new(StringComparer.Ordinal);
-        public string StatementPrefix { get; set; } = string.Empty;
-        public IdentifierScanState IdentifierScanState { get; set; }
-    }
 
     private const string ProcCallIdentifierPattern = @"(?:\[(?:[^\]\r\n]|\]\])+\]|`[^`\r\n]+`|""(?:""""|[^""\r\n])+""|##?\w+|[_\p{L}][\p{L}\p{Mn}\p{Mc}\p{Nd}\p{Pc}$]*(?:;\d+)?)";
     private const string ProcCallQualifierPattern = @"(?:(?:" + ProcCallIdentifierPattern + @")?\s*\.\s*)*";
